Move tests to new home
