+++ source/Energinet.DataHub.MeteringPoints.IntegrationTests/UpdateMasterData/EffectiveDateTests.cs
@@ -14,7 +14,10 @@
 
 using System.Threading.Tasks;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
+using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 using Energinet.DataHub.MeteringPoints.IntegrationTests.Tooling;
+using Energinet.DataHub.MeteringPoints.IntegrationTests.UpdateMasterData.ConsumptionMeteringPoints;
+using NodaTime.Text;
 using Xunit;
 using Xunit.Categories;
 
@@ -43,6 +46,39 @@ namespace Energinet.DataHub.MeteringPoints.IntegrationTests.UpdateMasterData
 
             AssertMasterData()
                 .HasEffectiveDate(EffectiveDate.Create(request.EffectiveDate));
+        }
+
+        [Fact]
+        public async Task Effective_date_is_required()
+        {
+            await SendCommandAsync(Scenarios.CreateVEProduction()
+                with
+                {
+                    EffectiveDate = string.Empty,
+                }).ConfigureAwait(false);
+
+            AssertValidationError("D02");
+        }
+
+        [Theory]
+        [InlineData("2021-01-01T18:00:00Z", "2021-01-02T22:00:00Z", true)]
+        [InlineData("2021-01-01T18:00:00Z", "2020-12-30T22:00:00Z", true)]
+        [InlineData("2021-01-01T18:00:00Z", "2021-01-01T22:00:00Z", false)]
+        [InlineData("2021-01-01T18:00:00Z", "2020-12-31T22:00:00Z", false)]
+        public async Task Effective_date_is_today_or_the_day_before(string today, string effectiveDate, bool expectError)
+        {
+            var timeProvider = GetService<ISystemDateTimeProvider>() as SystemDateTimeProviderStub;
+            timeProvider!.SetNow(InstantPattern.General.Parse(today).Value);
+
+            await SendCommandAsync(Scenarios.CreateVEProduction()).ConfigureAwait(false);
+
+            await SendCommandAsync(CreateUpdateRequest()
+                with
+                {
+                    EffectiveDate = effectiveDate,
+                }).ConfigureAwait(false);
+
+            AssertValidationError("E17", expectError);
         }
     }
 }
